Relocate protocolstvl_v2 (remove the v2)
